ReferralTextFieldWithCopy implemented
